+++ test/mvc_pattern_test.dart
@@ -20,29 +20,47 @@ void main() {
     /// You can directly access the 'internal workings' of the app!
     MyApp _app = tester.widget(find.byKey(key));
 
+    expect(_app, isInstanceOf<AppMVC>());
+
     /// Reference to the Controller.
     Controller _con = _app.controller;
+
+    expect(_con, isInstanceOf<ControllerMVC>());
 
     /// Reference to the StateMVC.
     StateMVC _sv = _con.stateMVC;
 
+    expect(_sv, isInstanceOf<State<StatefulWidget>>());
+
     /// The State object.
     State _state = _con.stateMVC;
+
+    expect(_state, isInstanceOf<State<StatefulWidget>>());
 
     /// Controller's unique identifier.
     String id = _con.keyId;
 
+    expect(id, isInstanceOf<String>());
+
     /// The StateView's unique identifier.
     String svId = _sv.keyId;
+
+    expect(svId, isInstanceOf<String>());
 
     /// Current context.
     BuildContext context2 = _sv.context;
 
+    expect(context2, isInstanceOf<BuildContext>());
+
     /// Is the widget mounted?
     bool mounted2 = _sv.mounted;
 
+    expect(mounted2, isInstanceOf<bool>());
+
     /// The StatefulWidget.
     StatefulWidget widget2 = _sv.widget;
+
+    expect(widget2, isInstanceOf<StatefulWidget>());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);